Task 6
This should fulfill the requirement left for this exercise.
If anything is still missing, please just write a mail.

Logging is the only feature still left, since only one example was
supposed to be implemented.
The structure in the source folder has no function, since the other
features are not implemented. The only reason it is still present is to
help understanding the pointcuts for the Logging-feature.